requirements.txt
# Deep Learning Framework
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Object Detection Libraries
ultralytics>=8.0.0
albumentations>=1.3.0
opencv-python>=4.8.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
lmdb>=1.4.0

# Visualization and Logging
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0
tqdm>=4.65.0

# Utilities
pyyaml>=6.0
pycocotools>=2.0.0
